requirements.txt
asgiref==3.5.2
Django==4.1.4
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Markdown==3.4.1
postgres==4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
psycopg2-pool==1.1
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
